refactor: campos de filtro movidos para um arquivo especifico e mudança de nomes em alguns lugares do código para melhor semantica.
