+++ src/feature/entities/IUser.ts
