+++ src/interfaces/index.ts
@@ -1,0 +1,5 @@
+export interface IFilterOptions {
+    checkbox: string
+    search: string
+    month: string
+}
